Robust Econometric Inference

Drawing statistical inference on the coefficients of a short- or long-horizon predictive regression with persistent regressors by using the IVX method of Magdalinos and Phillips (2009) DOI: 10.1017/S0266466608090154 and Kostakis, Magdalinos and Stamatogiannis (2015) DOI: 10.1093/rfs/hhu139.

See also

Author

Maintainer: Kostas Vasilopoulos k.vasilopoulo@gmail.com

Authors:

Kostas Vasilopoulos k.vasilopoulo@gmail.com

Efthymios Pavlidis e.pavlidis@lancaster.ac.uk