Fitter Functions for IVX Models

Basic function called by ivx to fit predictive models. These should only be used directly by experienced users.

ivx_fit(y, x, horizon = 1, offset = NULL, ...)

ivx_wfit(y, x, w, horizon = 1, offset = NULL, ...)

Arguments

y
vector of observations of length n, or a matrix with n rows.

x
design matrix of dimension n * p.

horizon
is the horizon (default horizon = 1 corresponds to a short-horizon regression).

offset
(numeric of length n). This can be used to specify an a priori known component to be included in the linear predictor during fitting.

...
currently disregarded.

w
vector of weights (length n) to be used in the fitting process for the wfit functions. Weighted least squares is used with weights w, i.e., sum(w * e^2) is minimized.

Examples

ivx_fit(monthly$Ret, as.matrix(monthly$LTY))

#> $coefficients
#>          x1 
#> -0.06649027 
#> 
#> $intercept
#> [1] 0.008216306
#> 
#> $fitted
#>    [1]  1.145873e-03  1.165821e-03  1.192417e-03  1.298801e-03  1.285503e-03
#>    [6]  1.325397e-03  1.278854e-03  1.285503e-03  1.312099e-03  1.305450e-03
#>   [11]  1.338695e-03  1.371940e-03  1.398536e-03  1.365291e-03  1.385238e-03
#>   [16]  1.391887e-03  1.378589e-03  1.325397e-03  1.332046e-03  1.212364e-03
#>   [21]  1.232311e-03  1.199066e-03  1.265556e-03  1.252258e-03  1.238960e-03
#>   [26]  1.179119e-03  1.086032e-03  9.929458e-04  1.119277e-03  1.019542e-03
#>   [31]  1.059436e-03  1.046138e-03  1.006244e-03  1.006244e-03  1.192417e-03
#>   [36]  1.298801e-03  1.238960e-03  1.192417e-03  1.245609e-03  1.272205e-03
#>   [41]  1.252258e-03  1.312099e-03  1.318748e-03  1.325397e-03  1.318748e-03
#>   [46]  1.345344e-03  1.345344e-03  1.358642e-03  1.305450e-03  1.219013e-03
#>   [51]  1.252258e-03  1.292152e-03  1.325397e-03  1.391887e-03  1.378589e-03
#>   [56]  1.338695e-03  1.332046e-03  1.152522e-03  9.397536e-04  9.397536e-04
#>   [61]  7.934750e-04  9.065085e-04  1.059436e-03  1.039489e-03  1.265556e-03
#>   [66]  1.179119e-03  1.192417e-03  1.371940e-03  1.365291e-03  1.378589e-03
#>   [71]  1.358642e-03  1.358642e-03  1.405185e-03  1.451729e-03  1.338695e-03
#>   [76]  1.365291e-03  1.338695e-03  1.451729e-03  1.465027e-03  1.445080e-03
#>   [81]  1.451729e-03  1.451729e-03  1.405185e-03  1.325397e-03  1.265556e-03
#>   [86]  1.365291e-03  1.391887e-03  1.458378e-03  1.504921e-03  1.558113e-03
#>   [91]  1.578060e-03  1.584709e-03  1.511570e-03  1.438431e-03  1.504921e-03
#>   [96]  1.511570e-03  1.551464e-03  1.631252e-03  1.671147e-03  1.677796e-03
#>  [101]  1.711041e-03  1.664498e-03  1.704392e-03  1.717690e-03  1.631252e-03
#>  [106]  1.624603e-03  1.644550e-03  1.637901e-03  1.664498e-03  1.604656e-03
#>  [111]  1.631252e-03  1.671147e-03  1.677796e-03  1.684445e-03  1.684445e-03
#>  [116]  1.697743e-03  1.744286e-03  1.717690e-03  1.711041e-03  1.790829e-03
#>  [121]  1.804127e-03  1.784180e-03  1.817425e-03  1.604656e-03  1.611305e-03
#>  [126]  1.624603e-03  1.604656e-03  1.657848e-03  1.598007e-03  1.611305e-03
#>  [131]  1.617954e-03  1.651199e-03  1.684445e-03  1.697743e-03  1.717690e-03
#>  [136]  1.684445e-03  1.777531e-03  1.790829e-03  1.777531e-03  1.790829e-03
#>  [141]  1.777531e-03  1.777531e-03  1.810776e-03  1.790829e-03  1.824074e-03
#>  [146]  1.844021e-03  1.870617e-03  1.923810e-03  1.977002e-03  2.056790e-03
#>  [151]  2.030194e-03  2.083386e-03  1.963704e-03  1.651199e-03  1.857319e-03
#>  [156]  1.930459e-03  1.996949e-03  1.977002e-03  1.983651e-03  2.070088e-03
#>  [161]  2.036843e-03  1.863968e-03  1.990300e-03  2.010247e-03  2.016896e-03
#>  [166]  2.070088e-03  2.076737e-03  2.176473e-03  2.209718e-03  2.083386e-03
#>  [171]  2.083386e-03  2.129929e-03  2.196420e-03  2.203069e-03  2.229665e-03
#>  [176]  2.229665e-03  2.236314e-03  2.216367e-03  2.289506e-03  2.262910e-03
#>  [181]  2.143227e-03  1.857319e-03  1.857319e-03  1.877266e-03  1.863968e-03
#>  [186]  1.883915e-03  1.877266e-03  1.877266e-03  1.877266e-03  1.877266e-03
#>  [191]  1.877266e-03  1.857319e-03  1.863968e-03  1.870617e-03  1.863968e-03
#>  [196]  1.857319e-03  1.863968e-03  1.877266e-03  1.877266e-03  1.870617e-03
#>  [201]  1.870617e-03  1.863968e-03  1.857319e-03  1.850670e-03  1.850670e-03
#>  [206]  1.850670e-03  1.857319e-03  1.857319e-03  1.850670e-03  1.857319e-03
#>  [211]  1.850670e-03  1.857319e-03  1.857319e-03  1.857319e-03  1.857319e-03
#>  [216]  1.857319e-03  1.863968e-03  1.903862e-03  1.930459e-03  1.930459e-03
#>  [221]  1.983651e-03  1.996949e-03  2.056790e-03  2.010247e-03  2.016896e-03
#>  [226]  2.030194e-03  2.063439e-03  2.103333e-03  2.176473e-03  2.176473e-03
#>  [231]  2.183122e-03  2.183122e-03  2.123280e-03  2.109982e-03  2.129929e-03
#>  [236]  2.109982e-03  2.056790e-03  2.043492e-03  2.063439e-03  2.036843e-03
#>  [241]  2.090035e-03  2.076737e-03  2.076737e-03  2.083386e-03  2.056790e-03
#>  [246]  2.063439e-03  2.063439e-03  2.076737e-03  2.103333e-03  2.083386e-03
#>  [251]  2.056790e-03  1.977002e-03  1.883915e-03  1.883915e-03  1.897213e-03
#>  [256]  1.897213e-03  1.910511e-03  1.963704e-03  1.917161e-03  1.897213e-03
#>  [261]  1.890564e-03  1.890564e-03  1.883915e-03  1.910511e-03  1.923810e-03
#>  [266]  1.950406e-03  1.963704e-03  1.990300e-03  1.990300e-03  1.990300e-03
#>  [271]  2.056790e-03  2.063439e-03  2.103333e-03  2.090035e-03  2.090035e-03
#>  [276]  2.090035e-03  2.109982e-03  2.070088e-03  2.076737e-03  2.070088e-03
#>  [281]  2.076737e-03  2.083386e-03  2.063439e-03  2.076737e-03  2.076737e-03
#>  [286]  2.036843e-03  2.003598e-03  2.010247e-03  2.010247e-03  2.030194e-03
#>  [291]  1.983651e-03  1.897213e-03  1.850670e-03  1.810776e-03  1.777531e-03
#>  [296]  1.824074e-03  1.863968e-03  1.817425e-03  1.810776e-03  1.744286e-03
#>  [301]  1.711041e-03  1.717690e-03  1.711041e-03  1.750935e-03  1.810776e-03
#>  [306]  1.790829e-03  1.777531e-03  1.764233e-03  1.724339e-03  1.657848e-03
#>  [311]  1.711041e-03  1.691094e-03  1.644550e-03  1.644550e-03  1.591358e-03
#>  [316]  1.544815e-03  1.484974e-03  1.411834e-03  1.498272e-03  1.498272e-03
#>  [321]  1.484974e-03  1.611305e-03  1.631252e-03  1.598007e-03  1.677796e-03
#>  [326]  1.564762e-03  1.644550e-03  1.651199e-03  1.684445e-03  1.644550e-03
#>  [331]  1.691094e-03  1.730988e-03  1.711041e-03  1.697743e-03  1.697743e-03
#>  [336]  1.677796e-03  1.691094e-03  1.598007e-03  1.558113e-03  1.584709e-03
#>  [341]  1.571411e-03  1.591358e-03  1.551464e-03  1.504921e-03  1.498272e-03
#>  [346]  1.518219e-03  1.558113e-03  1.538166e-03  1.538166e-03  1.558113e-03
#>  [351]  1.551464e-03  1.484974e-03  1.431782e-03  1.511570e-03  1.511570e-03
#>  [356]  1.418484e-03  1.338695e-03  1.345344e-03  1.312099e-03  1.285503e-03
#>  [361]  1.205715e-03  1.318748e-03  1.318748e-03  1.298801e-03  1.205715e-03
#>  [366]  1.185768e-03  1.099330e-03  1.072734e-03  1.059436e-03  1.079383e-03
#>  [371]  1.046138e-03  1.238960e-03  1.351993e-03  1.305450e-03  1.338695e-03
#>  [376]  1.365291e-03  1.431782e-03  1.418484e-03  1.345344e-03  1.219013e-03
#>  [381]  1.032840e-03  9.729987e-04  1.012893e-03  1.052787e-03  9.597007e-04
#>  [386]  7.868260e-04  8.267201e-04  8.200711e-04  7.469318e-04  7.269847e-04
#>  [391]  7.136867e-04  7.269847e-04  6.870906e-04  6.471964e-04  7.003886e-04
#>  [396]  6.272493e-04  5.275139e-04  5.674081e-04  6.471964e-04  7.668789e-04
#>  [401]  6.671435e-04  7.269847e-04  7.934750e-04  9.597007e-04  9.065085e-04
#>  [406]  9.264555e-04  8.998594e-04  8.466672e-04  9.729987e-04  8.134221e-04
#>  [411]  8.932104e-04  8.599653e-04  8.998594e-04  8.599653e-04  8.134221e-04
#>  [416]  8.134221e-04  7.735279e-04  8.200711e-04  8.400182e-04  8.134221e-04
#>  [421]  7.402828e-04  7.136867e-04  7.469318e-04  8.533162e-04  8.799124e-04
#>  [426]  8.865614e-04  8.333692e-04  7.602299e-04  8.333692e-04  8.533162e-04
#>  [431]  8.732633e-04  8.666143e-04  8.732633e-04  8.533162e-04  8.400182e-04
#>  [436]  8.333692e-04  8.067731e-04  8.001240e-04  7.934750e-04  7.934750e-04
#>  [441]  7.868260e-04  7.735279e-04  7.402828e-04  7.469318e-04  7.269847e-04
#>  [446]  7.003886e-04  6.804415e-04  6.804415e-04  6.870906e-04  6.937396e-04
#>  [451]  7.136867e-04  7.003886e-04  6.870906e-04  7.003886e-04  7.003886e-04
#>  [456]  6.937396e-04  6.870906e-04  6.937396e-04  6.804415e-04  6.937396e-04
#>  [461]  6.937396e-04  6.870906e-04  6.870906e-04  6.804415e-04  6.538454e-04
#>  [466]  6.206003e-04  6.206003e-04  5.674081e-04  5.075669e-04  4.610237e-04
#>  [471]  3.280431e-04  4.410766e-04  3.945334e-04  3.546392e-04  3.280431e-04
#>  [476]  2.947980e-04  1.817645e-04  3.080960e-04  3.945334e-04  3.080960e-04
#>  [481]  4.743217e-04  5.208649e-04  4.078315e-04  4.743217e-04  3.280431e-04
#>  [486]  2.947980e-04  1.285723e-04  1.418704e-04  8.202914e-05  6.208206e-05
#>  [491] -1.506868e-04 -2.703693e-04 -1.972300e-04 -6.424945e-05 -1.041436e-04
#>  [496] -2.238261e-04 -1.373887e-04 -1.373887e-04 -5.095140e-05  6.208206e-05
#>  [501]  4.213498e-05 -3.100432e-05 -1.107926e-04 -2.637203e-04 -4.764891e-04
#>  [506] -6.028206e-04 -6.094696e-04 -6.227677e-04 -4.432440e-04 -7.225031e-04
#>  [511] -6.427148e-04 -6.294167e-04 -6.892580e-04 -1.001762e-03 -8.421856e-04
#>  [516] -9.951132e-04 -1.068252e-03 -1.108147e-03 -8.288875e-04 -8.953778e-04
#>  [521] -1.148041e-03 -1.440598e-03 -1.214531e-03 -1.068252e-03 -1.114796e-03
#>  [526] -1.021709e-03 -1.108147e-03 -7.358012e-04 -8.089404e-04 -5.695755e-04
#>  [531] -6.826089e-04 -4.432440e-04 -6.161187e-04 -6.493638e-04 -7.623973e-04
#>  [536] -7.756953e-04 -5.562774e-04 -4.764891e-04 -4.099988e-04 -4.631911e-04
#>  [541] -4.698401e-04 -5.296813e-04 -5.030852e-04 -5.762245e-04 -5.895226e-04
#>  [546] -4.698401e-04 -5.363303e-04 -4.432440e-04 -4.565420e-04 -5.296813e-04
#>  [551] -4.299459e-04 -3.368595e-04 -4.831381e-04 -1.054954e-03 -1.074902e-03
#>  [556] -1.061603e-03 -1.068252e-03 -1.174637e-03 -1.221180e-03 -1.553631e-03
#>  [561] -1.340863e-03 -1.174637e-03 -1.081551e-03 -1.234478e-03 -1.327565e-03
#>  [566] -1.420651e-03 -1.473843e-03 -1.706559e-03 -1.925977e-03 -1.886083e-03
#>  [571] -1.899381e-03 -1.972520e-03 -2.185289e-03 -2.065607e-03 -1.786347e-03
#>  [576] -1.626771e-03 -1.553631e-03 -1.792996e-03 -1.739804e-03 -1.979169e-03
#>  [581] -2.165342e-03 -2.058957e-03 -1.906030e-03 -2.012414e-03 -2.112150e-03
#>  [586] -2.231832e-03 -1.945924e-03 -2.072256e-03 -1.852838e-03 -1.832891e-03
#>  [591] -1.832891e-03 -1.766400e-03 -1.799645e-03 -1.959222e-03 -1.866136e-03
#>  [596] -1.852838e-03 -1.753102e-03 -1.693261e-03 -1.679963e-03 -1.480492e-03
#>  [601] -1.294319e-03 -1.580228e-03 -1.653367e-03 -1.633420e-03 -1.626771e-03
#>  [606] -1.586877e-03 -1.513737e-03 -1.606824e-03 -1.513737e-03 -1.580228e-03
#>  [611] -1.693261e-03 -1.666665e-03 -1.839540e-03 -1.925977e-03 -1.965871e-03
#>  [616] -2.025712e-03 -2.072256e-03 -2.165342e-03 -2.251779e-03 -2.205236e-03
#>  [621] -2.105501e-03 -2.218534e-03 -2.411356e-03 -2.331568e-03 -2.471197e-03
#>  [626] -2.391409e-03 -2.537687e-03 -2.497793e-03 -2.630774e-03 -2.504442e-03
#>  [631] -2.331568e-03 -2.451250e-03 -2.531038e-03 -2.664019e-03 -3.375465e-03
#>  [636] -3.209239e-03 -3.229186e-03 -3.907387e-03 -4.386117e-03 -4.738515e-03
#>  [641] -3.654724e-03 -3.395412e-03 -3.189292e-03 -3.641426e-03 -4.080262e-03
#>  [646] -4.379468e-03 -4.685323e-03 -4.678674e-03 -4.472554e-03 -4.552343e-03
#>  [651] -5.031072e-03 -4.798356e-03 -5.356875e-03 -4.911390e-03 -5.170702e-03
#>  [656] -5.609538e-03 -6.108215e-03 -6.354229e-03 -5.702624e-03 -4.612184e-03
#>  [661] -5.370173e-03 -5.908744e-03 -5.822307e-03 -5.722571e-03 -5.463259e-03
#>  [666] -5.529749e-03 -5.888797e-03 -5.489855e-03 -4.838251e-03 -4.366170e-03
#>  [671] -3.894089e-03 -3.980526e-03 -3.781055e-03 -3.900738e-03 -3.548339e-03
#>  [676] -3.701267e-03 -3.488498e-03 -3.894089e-03 -3.940632e-03 -4.465905e-03
#>  [681] -4.545693e-03 -4.193295e-03 -4.399415e-03 -4.319627e-03 -4.459256e-03
#>  [686] -4.346223e-03 -4.592237e-03 -4.831602e-03 -5.037721e-03 -5.682677e-03
#>  [691] -5.636134e-03 -5.097563e-03 -4.944635e-03 -4.711919e-03 -4.299679e-03
#>  [696] -4.273083e-03 -4.279732e-03 -3.993824e-03 -4.539044e-03 -4.352872e-03
#>  [701] -4.226540e-03 -3.561638e-03 -3.515094e-03 -3.754459e-03 -3.601532e-03
#>  [706] -3.694618e-03 -3.488498e-03 -3.222537e-03 -2.856841e-03 -2.870139e-03
#>  [711] -2.092203e-03 -1.593526e-03 -1.699910e-03 -2.138746e-03 -1.753102e-03
#>  [716] -1.879434e-03 -1.573579e-03 -1.999116e-03 -1.839540e-03 -1.679963e-03
#>  [721] -1.746453e-03 -1.673314e-03 -1.573579e-03 -1.786347e-03 -2.211885e-03
#>  [726] -2.351515e-03 -2.331568e-03 -2.531038e-03 -2.723860e-03 -3.096206e-03
#>  [731] -2.657370e-03 -2.690615e-03 -2.617476e-03 -2.165342e-03 -2.178640e-03
#>  [736] -2.491144e-03 -2.677317e-03 -2.830245e-03 -2.597529e-03 -2.796999e-03
#>  [741] -2.816947e-03 -2.597529e-03 -2.411356e-03 -2.637423e-03 -2.604178e-03
#>  [746] -2.504442e-03 -2.717211e-03 -2.677317e-03 -2.604178e-03 -2.338217e-03
#>  [751] -1.959222e-03 -1.826242e-03 -2.092203e-03 -2.132097e-03 -1.886083e-03
#>  [756] -1.872785e-03 -1.925977e-03 -2.251779e-03 -2.324919e-03 -2.411356e-03
#>  [761] -2.644072e-03 -2.371462e-03 -2.245130e-03 -2.218534e-03 -2.617476e-03
#>  [766] -2.577582e-03 -2.471197e-03 -2.205236e-03 -2.112150e-03 -2.065607e-03
#>  [771] -2.092203e-03 -2.112150e-03 -2.065607e-03 -2.118799e-03 -2.218534e-03
#>  [776] -2.152044e-03 -1.939275e-03 -1.753102e-03 -1.759751e-03 -1.746453e-03
#>  [781] -1.354161e-03 -1.660016e-03 -1.666665e-03 -1.799645e-03 -1.839540e-03
#>  [786] -1.693261e-03 -1.586877e-03 -1.327565e-03 -1.320916e-03 -1.221180e-03
#>  [791] -1.427300e-03 -1.473843e-03 -1.327565e-03 -1.320916e-03 -1.141392e-03
#>  [796] -1.167988e-03 -1.161339e-03 -1.161339e-03 -9.419210e-04 -8.621327e-04
#>  [801] -6.427148e-04 -6.693109e-04 -6.427148e-04 -8.288875e-04 -8.488346e-04
#>  [806] -7.358012e-04 -1.035007e-03 -1.320916e-03 -1.453896e-03 -1.546982e-03
#>  [811] -1.646718e-03 -1.460545e-03 -1.560280e-03 -1.819593e-03 -1.879434e-03
#>  [816] -1.872785e-03 -1.812943e-03 -1.686612e-03 -1.540333e-03 -1.520386e-03
#>  [821] -1.453896e-03 -1.001762e-03 -9.552190e-04 -1.094849e-03 -9.818151e-04
#>  [826] -9.086758e-04 -7.623973e-04 -6.427148e-04 -5.097342e-04 -5.496284e-04
#>  [831] -8.820797e-04 -1.048305e-03 -1.194584e-03 -1.267723e-03 -1.174637e-03
#>  [836] -1.201233e-03 -1.327565e-03 -1.181286e-03 -9.618681e-04 -7.756953e-04
#>  [841] -9.751661e-04 -1.081551e-03 -1.114796e-03 -1.307617e-03 -1.187935e-03
#>  [846] -1.161339e-03 -1.074902e-03 -7.358012e-04 -9.685171e-04 -8.155895e-04
#>  [851] -6.427148e-04 -5.828735e-04 -5.030852e-04 -4.166479e-04 -4.831381e-04
#>  [856] -5.030852e-04 -5.163833e-04 -4.365949e-04 -3.302105e-04 -3.834027e-04
#>  [861] -1.373887e-04  6.208206e-05 -9.084556e-05 -5.760042e-05 -1.041436e-04
#>  [866] -6.424945e-05 -4.033498e-04 -4.365949e-04 -4.498930e-04 -5.895226e-04
#>  [871] -6.693109e-04 -7.490992e-04 -8.155895e-04 -7.956424e-04 -8.288875e-04
#>  [876] -9.020268e-04 -1.035007e-03 -9.286229e-04 -7.956424e-04 -6.094696e-04
#>  [881] -6.892580e-04 -7.557482e-04 -6.360657e-04 -5.629265e-04 -4.498930e-04
#>  [886] -5.695755e-04 -4.897872e-04 -3.302105e-04 -2.105280e-04 -2.371241e-04
#>  [891] -1.506868e-04 -2.171771e-04 -4.432440e-04 -4.498930e-04 -4.232969e-04
#>  [896] -2.304751e-04 -1.307397e-04 -1.041436e-04  1.352214e-04 -1.772829e-04
#>  [901] -3.235615e-04 -2.836673e-04 -2.437732e-04 -5.163833e-04 -3.235615e-04
#>  [906] -3.435086e-04 -2.637203e-04 -1.174417e-04  1.086252e-04  3.080960e-04
#>  [911]  1.219233e-04  3.548595e-05  2.814999e-04  2.083606e-04  3.612883e-04
#>  [916]  2.682019e-04  3.014470e-04  6.006532e-04  4.942688e-04 -1.041436e-04
#>  [921] -3.765334e-05  2.416058e-04  5.543303e-05  4.878401e-05  1.019762e-04
#>  [926]  1.817645e-04  2.881490e-04  3.479902e-04 -3.100432e-05 -8.419653e-05
#>  [931] -3.765334e-05  2.218790e-05  2.216587e-04  2.549038e-04  3.213941e-04
#>  [936]  1.618175e-04  2.814999e-04  4.078315e-04  3.147451e-04  2.549038e-04
#>  [941]  4.344276e-04  5.740571e-04  6.471964e-04  4.676727e-04  6.272493e-04
#>  [946]  4.144805e-04  2.814999e-04  3.014470e-04  4.344276e-04  3.479902e-04
#>  [951]  4.610237e-04  1.285723e-04 -3.765334e-05 -5.760042e-05 -3.100432e-05
#>  [956]  5.543303e-05  2.017116e-04  2.814999e-04  3.014470e-04  3.945334e-04
#>  [961]  2.349568e-04  1.618175e-04  3.280431e-04  2.216587e-04  2.482548e-04
#>  [966]  1.086252e-04  3.548595e-05  1.684665e-04  2.615529e-04  2.482548e-04
#>  [971]  3.080960e-04  5.408120e-04  5.075669e-04  6.006532e-04  5.873552e-04
#>  [976]  6.272493e-04  4.543746e-04  3.413412e-04  4.410766e-04  4.078315e-04
#>  [981]  5.142159e-04  5.541100e-04  3.213941e-04  1.026191e-03  1.484974e-03
#>  [986]  8.799124e-04  8.333692e-04  1.139224e-03  7.735279e-04  6.272493e-04
#>  [991]  6.471964e-04  6.405474e-04  7.402828e-04  8.200711e-04  7.070377e-04
#>  [996]  8.001240e-04  4.543746e-04  5.701408e-04  5.693230e-04  4.558574e-04
#> [1001]  5.916238e-04  7.947383e-04  9.991028e-04  9.957650e-04  1.324879e-03
#> [1006]  1.234771e-03  1.058392e-03  9.712966e-04  7.483813e-04  6.272493e-04
#> [1011]  6.671435e-04  6.471964e-04  7.336338e-04  8.998594e-04  8.134221e-04
#> [1016]  1.066085e-03  1.411834e-03  1.737637e-03  1.584709e-03  1.697743e-03
#> [1021]  1.850670e-03  1.844021e-03  1.711041e-03  1.571411e-03  1.764233e-03
#> [1026]  2.043492e-03  2.003598e-03  2.129929e-03  2.050141e-03  1.996949e-03
#> [1031]  1.930459e-03  2.043492e-03
#> 
#> $residuals
#>    [1] -1.106232e-02  3.617293e-02 -1.319964e-03  4.523914e-03  5.245894e-02
#>    [6] -2.939202e-02  7.256124e-02  2.446436e-02  4.337048e-02 -5.635967e-02
#>   [11]  5.946780e-02  1.045238e-02 -1.617011e-02 -2.106128e-02  9.189176e-02
#>   [16]  2.548305e-02  3.449476e-03 -4.423619e-02  1.179727e-03  5.853946e-02
#>   [21]  1.195596e-02  4.752044e-03  1.056874e-01 -3.260299e-03  4.333946e-02
#>   [26] -1.315554e-02 -7.852788e-03  9.653226e-03 -5.197149e-02  9.480776e-02
#>   [31]  3.877879e-02  8.181815e-02 -6.214595e-02 -2.133514e-01 -1.430003e-01
#>   [36]  1.834080e-02  4.880301e-02  1.661131e-02  6.137757e-02 -2.261857e-02
#>   [41] -1.948565e-02 -1.767575e-01  3.031689e-02  3.608436e-03 -1.389218e-01
#>   [46] -9.513020e-02 -2.810303e-02 -8.167427e-02  5.083380e-02  9.823585e-02
#>   [51] -7.919815e-02 -1.043550e-01 -1.549508e-01  1.300773e-01 -8.461380e-02
#>   [56]  1.186152e-03 -3.451365e-01  6.714019e-02 -1.147339e-01 -1.562516e-01
#>   [61] -2.700767e-02  4.403688e-02 -1.350573e-01 -2.401410e-01 -2.720773e-01
#>   [66] -1.722025e-03  3.258759e-01  3.326631e-01 -3.566928e-02 -1.607626e-01
#>   [71] -5.950024e-02  5.304979e-02  8.052460e-03 -2.040498e-01  3.326322e-02
#>   [76]  3.410280e-01  1.553463e-01  1.235976e-01 -9.463364e-02  1.112806e-01
#>   [81] -1.318058e-01 -1.005922e-01  9.741203e-02  2.154860e-02  1.023782e-01
#>   [86] -4.018861e-02 -1.079578e-03 -3.252044e-02 -8.857497e-02  1.713087e-02
#>   [91] -1.341839e-01  5.093058e-02 -6.855849e-03 -3.950606e-02  7.677006e-02
#>   [96] -6.458723e-03 -4.978892e-02 -4.346879e-02 -3.931490e-02  8.530497e-02
#>  [101]  3.030409e-02  6.143277e-02  7.407233e-02  1.968255e-02  1.961453e-02
#>  [106]  6.323917e-02  4.439968e-02  3.272749e-02  6.272505e-02  1.658791e-02
#>  [111]  1.826411e-02 -8.753614e-02  4.505046e-02  2.562286e-02  6.435470e-02
#>  [116]  7.436528e-03 -1.733696e-03  6.713439e-02  3.963643e-03 -1.050300e-02
#>  [121]  2.947042e-02  1.193204e-02 -9.780173e-03 -9.063436e-02 -1.233712e-02
#>  [126] -6.234997e-02  8.762050e-02 -5.911130e-02 -1.571820e-01 -1.154895e-01
#>  [131] -9.505877e-02 -5.315707e-02  6.166783e-03  5.741828e-02 -3.001317e-01
#>  [136]  1.267678e-01 -4.760456e-02  2.199805e-01  6.592527e-02 -3.274820e-02
#>  [141]  1.032285e-02  6.875630e-02 -3.285106e-02  3.764594e-02 -7.992954e-02
#>  [146]  3.177836e-02 -1.510998e-01 -1.360277e-02  6.028175e-02 -7.081355e-02
#>  [151]  9.865028e-02 -7.663506e-02  1.610549e-01 -2.381630e-02 -4.790214e-02
#>  [156]  2.108422e-02 -4.351648e-02  6.377127e-03  6.852574e-03 -1.133190e-02
#>  [161] -2.698232e-01  7.284496e-02  2.592599e-02  2.824753e-02  7.654668e-03
#>  [166]  3.396268e-02 -3.615715e-02 -4.186648e-03 -5.364678e-02 -1.135324e-02
#>  [171]  2.375059e-03 -7.083608e-02  6.577088e-03  5.028322e-02  4.836352e-02
#>  [176] -8.253247e-03 -1.419901e-02 -7.322474e-02 -3.187046e-02 -4.664291e-02
#>  [181]  1.375655e-02 -2.516403e-02 -7.283790e-02 -4.567903e-02  6.176415e-02
#>  [186]  1.631066e-02  2.857341e-02  8.190914e-03  2.506992e-02  5.830483e-02
#>  [191] -1.038982e-02  4.864980e-02  6.273978e-02  4.972191e-02  5.037471e-02
#>  [196] -3.314123e-03  4.427166e-02  1.478653e-02 -6.095394e-02  8.998070e-03
#>  [201]  2.027174e-02 -1.842884e-02 -7.570730e-02  6.087837e-02  1.135089e-02
#>  [206] -4.448009e-04  1.395869e-02 -1.922979e-02  4.064980e-02  4.580211e-02
#>  [211] -2.661074e-02  7.458017e-03 -7.534994e-03 -4.143509e-03  1.032152e-02
#>  [216]  3.686552e-02  5.260562e-03  6.013715e-02 -4.992536e-02  7.967830e-02
#>  [221]  1.290053e-02 -5.395716e-03 -2.856656e-02  5.055323e-02  3.801645e-02
#>  [226]  2.257911e-02  3.682115e-02  3.006753e-03  6.151930e-02 -7.480169e-02
#>  [231]  3.940243e-02  2.820035e-02  2.212093e-02 -4.331021e-02 -3.495547e-02
#>  [236] -7.803285e-02 -1.152732e-01 -1.194084e-02 -9.695311e-03  3.915062e-02
#>  [241]  1.466868e-02 -1.543589e-02 -2.221697e-02 -4.746773e-02 -8.066739e-03
#>  [246]  4.550491e-02  3.172916e-02 -3.067487e-02 -1.740760e-02  1.291748e-02
#>  [251] -2.526010e-02  2.287740e-02 -4.584878e-02 -4.845541e-02  7.086337e-02
#>  [256]  2.292162e-02  7.323109e-02  1.382428e-04 -6.111340e-02  7.963848e-03
#>  [261] -3.482705e-02  6.012381e-02 -1.110432e-01  2.776184e-02 -4.919421e-03
#>  [266] -3.977713e-02  2.577830e-02 -2.653355e-02 -3.753123e-02 -5.923184e-03
#>  [271]  5.311080e-02  1.395088e-02  1.937450e-02  2.166304e-02  1.286236e-02
#>  [276]  4.203735e-02  1.178434e-02  1.175212e-02  4.412090e-04  3.716150e-02
#>  [281]  3.809892e-02 -6.349995e-02  9.116281e-03  3.919992e-02  4.998793e-02
#>  [286] -2.749310e-03  1.438664e-02  4.609975e-02  5.524625e-02  6.594144e-03
#>  [291] -2.545583e-02  4.017624e-02 -3.821332e-02 -3.399038e-02  6.084519e-02
#>  [296]  4.117560e-02 -5.050003e-03 -2.128620e-02  3.529518e-03  3.322010e-02
#>  [301]  9.759485e-03 -3.329103e-02  4.376364e-02 -5.008128e-02  2.532357e-02
#>  [306]  3.884090e-02  1.105431e-02 -1.366243e-02 -2.619955e-02 -7.462726e-03
#>  [311]  5.171100e-02  2.957048e-02 -1.431189e-02 -1.554596e-02 -3.047265e-02
#>  [316] -3.355301e-02 -6.354038e-04 -2.227979e-02  1.801784e-02 -5.929728e-02
#>  [321] -5.699759e-03  4.198316e-02  1.600460e-02 -5.786594e-03  4.439358e-02
#>  [326]  5.584847e-03  2.291984e-02  4.001513e-02  3.469406e-02 -4.514562e-03
#>  [331]  5.011756e-02 -3.326680e-02  7.209493e-02 -2.526844e-02  8.424605e-02
#>  [336]  4.571232e-02  1.056476e-02  6.120342e-03 -9.477995e-03  3.171809e-02
#>  [341] -1.183978e-03  7.066422e-02  4.861907e-02 -8.186416e-03  3.441595e-03
#>  [346] -3.888437e-02  7.606473e-02 -8.412981e-03 -4.403968e-02  3.229781e-02
#>  [351]  5.978137e-02 -8.696261e-03 -6.848654e-02  2.978704e-02  4.186303e-02
#>  [356] -4.138005e-02 -5.322424e-02 -1.812864e-03 -1.034954e-02  2.716633e-02
#>  [361] -4.964793e-02 -3.384264e-02  1.491965e-02  3.680128e-02  2.943995e-02
#>  [366] -9.213667e-03  1.091402e-03 -5.856685e-02 -6.822339e-02 -4.812494e-02
#>  [371]  1.455208e-02 -4.693841e-02  3.627759e-02 -2.209896e-02  2.526963e-02
#>  [376]  2.575191e-02  1.401031e-02  2.075920e-02  3.636019e-02  1.105199e-02
#>  [381]  4.238887e-02  1.963765e-02  2.023149e-02  4.447871e-02 -1.564747e-03
#>  [386] -1.064878e-03 -4.596091e-03  3.146918e-02  1.533481e-02 -8.653066e-03
#>  [391]  2.566212e-02 -1.920750e-02 -5.221940e-02  3.715446e-03  9.440619e-03
#>  [396]  2.036564e-02 -8.157997e-02  4.688741e-03 -2.131253e-02 -2.452848e-02
#>  [401]  2.478669e-02  1.364370e-02 -3.075967e-02  2.345710e-02 -6.795501e-02
#>  [406] -8.882539e-03  3.738709e-02  4.007937e-02  5.470821e-02  2.489824e-02
#>  [411]  2.128485e-02 -2.994884e-03  1.568378e-02 -3.450558e-02  2.574666e-02
#>  [416]  1.708803e-02 -2.632379e-02  2.048493e-02  3.669611e-02 -2.856962e-03
#>  [421] -4.571757e-02  1.268561e-02 -1.155051e-02 -7.157522e-02 -9.115109e-02
#>  [426] -9.195954e-02  5.543293e-02  1.354612e-02 -5.792921e-02  4.104934e-04
#>  [431]  9.494360e-02  7.491571e-03  4.117671e-02 -3.246258e-02  2.795656e-02
#>  [436]  3.996441e-02  1.156160e-02 -2.671991e-02 -9.805543e-03  4.394487e-02
#>  [441] -1.870699e-02  2.594525e-02 -1.280909e-02  1.735429e-02  1.866571e-02
#>  [446]  8.541598e-03  8.766583e-03 -2.154031e-03  8.399768e-03  9.632046e-03
#>  [451]  1.042310e-02 -1.996946e-02  2.164575e-02  3.684066e-04 -6.775088e-03
#>  [456] -3.208941e-03  2.561665e-02 -4.825549e-03 -2.093912e-02  2.632833e-02
#>  [461] -1.179111e-02 -5.663882e-02  5.832486e-03  1.861089e-02  2.427186e-02
#>  [466]  1.982196e-02 -1.250247e-02  1.625029e-03 -3.162974e-03 -2.181155e-02
#>  [471] -3.220699e-02  1.295646e-02 -5.932642e-02 -2.339908e-02 -2.097503e-02
#>  [476] -8.333412e-02 -1.527099e-02  3.772733e-02  1.616691e-03 -9.106525e-03
#>  [481]  6.693395e-02 -1.661413e-03  3.141192e-02  3.373794e-02 -5.616480e-02
#>  [486]  1.114361e-02  3.763224e-02 -1.519229e-02  2.492096e-02 -3.627122e-02
#>  [491] -1.928568e-03  1.835446e-02 -5.141375e-02 -3.606303e-02  9.111862e-04
#>  [496]  7.196668e-02  7.204627e-03  5.981295e-04 -2.683728e-02  7.103569e-03
#>  [501]  2.962081e-02 -4.351686e-04  4.315542e-02 -5.065129e-02 -1.675435e-02
#>  [506] -5.244427e-02  2.507518e-02  1.343341e-02 -5.489752e-03 -6.488921e-02
#>  [511] -6.983159e-02  3.482548e-02 -3.404364e-02  3.624339e-02 -3.926040e-02
#>  [516] -2.678775e-02 -8.852002e-02  4.768730e-02 -6.993967e-03 -1.023421e-01
#>  [521] -6.520444e-02 -5.921892e-02  6.376952e-02  3.958552e-02  2.519997e-02
#>  [526] -1.783921e-02  4.445395e-02  4.867124e-02  3.399423e-02  5.404672e-03
#>  [531]  3.122734e-02  2.922787e-02 -4.515387e-02 -5.059508e-03 -4.957279e-02
#>  [536]  3.276144e-02 -1.330351e-02 -4.908145e-02 -5.106810e-03  7.648270e-02
#>  [541]  1.134101e-02  2.278948e-02  4.275026e-04 -1.767067e-03  1.458187e-02
#>  [546] -2.766319e-02 -4.177800e-03  3.109890e-02 -1.087108e-02  2.772289e-03
#>  [551]  4.085987e-02  5.231101e-03 -2.532770e-02 -4.254714e-02 -9.375348e-03
#>  [556] -4.959007e-02 -2.288391e-02 -1.400366e-02  2.892845e-02 -4.185226e-02
#>  [561]  3.079309e-02 -1.009883e-02 -1.230330e-01  7.841264e-03 -1.771556e-02
#>  [566] -7.744298e-03 -3.120233e-02 -4.798174e-02 -3.695686e-02 -2.254851e-02
#>  [571] -8.916212e-02 -9.475099e-02 -1.349634e-01  1.460602e-01 -5.587653e-02
#>  [576] -2.746513e-02  1.074589e-01  5.717104e-02  1.619630e-02  4.095520e-02
#>  [581]  4.272611e-02  3.819292e-02 -7.498287e-02 -2.357072e-02 -4.072250e-02
#>  [586]  5.503389e-02  2.193592e-02 -1.772521e-02  1.060853e-01 -1.269070e-02
#>  [591]  2.519226e-02 -1.674892e-02 -1.509213e-02  3.595463e-02 -1.456097e-02
#>  [596] -5.599143e-03  1.690565e-02 -2.765001e-02 -8.154037e-03  4.520809e-02
#>  [601] -5.795412e-02 -2.272478e-02 -1.951239e-02 -6.252631e-03 -2.202888e-02
#>  [606]  3.943104e-02 -2.293535e-02 -1.979609e-02 -8.577010e-03 -5.084140e-02
#>  [611]  2.796497e-02 -2.550087e-03 -7.046656e-02 -2.452083e-02  1.847332e-02
#>  [616]  7.732368e-02  5.769185e-03 -2.457990e-02  4.736838e-02  2.476725e-02
#>  [621] -1.365438e-02 -1.028409e-01  1.749318e-02  6.946808e-03  3.204280e-02
#>  [626] -3.913829e-02  4.587637e-02 -5.897730e-03 -2.688352e-02  2.935937e-02
#>  [631]  2.088051e-03  5.010670e-02 -8.656833e-03 -7.835834e-02  4.040328e-02
#>  [636]  6.905201e-03  4.804712e-02 -8.136187e-03 -1.135809e-01  2.897055e-02
#>  [641]  4.208538e-02  2.085887e-02  5.858906e-02  5.610674e-03  1.935182e-02
#>  [646]  8.808987e-03  9.462666e-02 -4.251190e-02 -5.739260e-02  8.456535e-03
#>  [651]  2.565088e-02 -3.139233e-02 -4.170001e-03 -2.157062e-02 -1.101667e-02
#>  [656] -6.861461e-02 -6.321875e-02  4.028024e-02  3.267540e-02 -3.652979e-02
#>  [661] -2.381800e-02 -6.073104e-02 -1.687750e-02  3.122049e-02 -3.971433e-02
#>  [666] -2.510049e-02 -2.955588e-02  1.112910e-01  3.371412e-03  1.009212e-01
#>  [671]  3.752696e-02  9.272060e-03  2.716982e-02  1.807873e-02  2.696733e-02
#>  [676]  6.505547e-02 -1.373394e-02  2.869443e-02 -3.979482e-02  7.966364e-03
#>  [681]  6.144449e-03 -2.117423e-02  1.609050e-02 -1.342029e-02 -1.161922e-02
#>  [686] -4.237056e-02  9.492086e-03 -1.013156e-03 -6.230022e-02  1.550144e-02
#>  [691] -2.111668e-02  9.855275e-02 -8.507364e-03 -5.267998e-03 -1.837481e-02
#>  [696]  1.765777e-02  6.694151e-02  6.913736e-03 -7.666026e-03 -1.113376e-02
#>  [701]  5.377303e-02  8.584134e-03 -1.028723e-02 -1.332308e-02 -4.084496e-02
#>  [706]  3.718703e-02  6.225937e-02  3.953892e-02 -3.048270e-03  6.603318e-02
#>  [711]  4.585435e-02 -2.199191e-02  4.581150e-02  7.947343e-03 -6.737226e-02
#>  [716]  6.411021e-02 -9.463785e-02  4.796353e-02  1.740295e-02 -3.536787e-02
#>  [721]  1.192417e-01  3.338655e-02  1.822040e-02 -1.664816e-02  1.908647e-03
#>  [726]  4.186858e-02  4.155748e-02  3.146239e-02 -2.918195e-02 -2.500086e-01
#>  [731] -9.255944e-02  6.457668e-02  3.494484e-02  3.898838e-02 -3.856056e-02
#>  [736]  3.512641e-03  1.043639e-03  3.950068e-02 -1.194017e-02 -4.052396e-02
#>  [741]  3.410092e-02  1.841727e-02 -2.285483e-02  9.873728e-03  6.100952e-02
#>  [746] -3.441403e-02  1.370981e-02  4.106791e-02  3.092378e-02 -1.498986e-02
#>  [751]  7.671599e-02  1.023870e-02 -1.318243e-02 -3.292305e-02  1.111088e-02
#>  [756]  1.318820e-02 -7.917186e-02  4.496719e-03  1.743547e-02 -3.366046e-02
#>  [761]  8.445016e-02 -1.553558e-02 -1.334067e-02 -1.040035e-01 -5.847061e-02
#>  [766] -1.184134e-02  5.437564e-02  1.924082e-02  3.582820e-02  6.131502e-02
#>  [771]  1.651574e-02 -5.236171e-03  3.470154e-02 -5.353955e-02  3.836287e-02
#>  [776]  1.641620e-02 -2.419816e-02  6.079941e-03 -4.813574e-02  1.012898e-01
#>  [781] -2.525102e-02  6.580461e-03 -2.589110e-02  2.195006e-02 -1.143151e-03
#>  [786] -2.117211e-02  3.394946e-02 -2.687516e-02  5.291322e-03 -1.117189e-03
#>  [791]  2.758149e-02  7.459344e-03  1.202599e-03  7.558535e-03  1.498223e-02
#>  [796] -3.081277e-02  1.999184e-02 -2.543124e-03 -1.091757e-02  3.037863e-02
#>  [801] -1.418252e-02  1.320205e-02 -1.526430e-02  6.111790e-03  2.648364e-02
#>  [806] -3.398197e-02 -5.079900e-02  6.443211e-03  9.868518e-03 -3.167166e-02
#>  [811]  2.569210e-02  3.366717e-02 -3.109816e-02  1.564313e-02 -4.385920e-02
#>  [816]  7.617576e-03  1.812404e-02  3.071685e-02  2.095467e-02  2.102716e-02
#>  [821]  3.109281e-02  1.545493e-02  2.456558e-02 -4.623582e-03  3.303090e-02
#>  [826] -1.163045e-02  3.475848e-02  8.974323e-03  2.580183e-02  1.727244e-03
#>  [831]  1.703673e-03  7.242289e-03  1.732054e-02 -3.488221e-03 -5.321767e-02
#>  [836]  1.390895e-02  4.694855e-02  1.869297e-02  6.578455e-02 -2.812200e-02
#>  [841]  5.307818e-02 -8.008824e-05 -5.029771e-02  5.085996e-02  5.178605e-02
#>  [846]  3.541914e-02  6.959438e-02 -6.470315e-02  4.447778e-02 -4.115015e-02
#>  [851]  3.708896e-02  8.609497e-03  3.434105e-03  6.100962e-02  4.161134e-02
#>  [856]  2.831250e-03 -2.613302e-02  3.173072e-02 -1.867249e-02 -1.629043e-01
#>  [861]  5.262218e-02  6.846658e-02  5.240472e-02  4.964887e-02  3.366105e-02
#>  [866] -4.070278e-02  3.024784e-02  2.891342e-02 -3.125403e-02  4.514698e-02
#>  [871] -3.851867e-02 -1.273772e-02 -3.629162e-02  5.446144e-02  1.269441e-02
#>  [876]  5.268406e-02 -5.886845e-02 -2.582015e-02  8.541879e-02 -4.091640e-02
#>  [881] -3.138413e-02  1.712269e-02 -2.086579e-02  5.266299e-02 -6.265376e-02
#>  [886] -1.334472e-02 -9.084156e-02 -3.851103e-03  2.245570e-02 -1.042741e-01
#>  [891] -7.421326e-02  6.673015e-02 -4.690234e-04 -3.198311e-02 -1.655545e-02
#>  [896] -7.292745e-02 -9.121095e-02  1.257492e-02  6.920375e-02  2.928616e-03
#>  [901] -2.020916e-02 -2.546448e-02  3.098664e-02 -6.863718e-02 -1.364017e-02
#>  [906] -7.981701e-02 -8.270683e-02  1.035894e-03 -1.215907e-01  7.852520e-02
#>  [911]  5.120917e-02 -6.639655e-02 -3.291212e-02 -2.124714e-02  4.218581e-03
#>  [916]  7.361600e-02  4.593806e-02  6.541807e-03  1.181164e-02  1.406104e-02
#>  [921] -1.624960e-02  4.834596e-02  3.694743e-03  4.501859e-02  1.334576e-02
#>  [926]  8.684593e-03 -2.088849e-02 -2.153856e-02  8.164992e-03  1.383622e-02
#>  [931] -3.920743e-02 -2.030976e-03  4.603842e-03  8.532211e-03  3.327409e-02
#>  [936]  2.659132e-02 -3.091806e-02  1.388197e-02 -2.452035e-02 -2.634386e-02
#>  [941]  2.404360e-02 -5.923240e-03  2.887254e-02 -1.704945e-02 -2.214935e-04
#>  [946] -2.363547e-02  2.957812e-02 -8.290406e-03  1.771117e-02 -6.152874e-03
#>  [951]  3.979656e-03  3.404192e-03 -3.721588e-02 -7.012070e-03 -3.118902e-03
#>  [956]  1.500974e-02  1.711493e-02  2.301394e-02  9.124134e-03  4.461614e-03
#>  [961]  6.264063e-03 -2.846114e-02  1.625277e-03  3.402015e-02  2.487937e-02
#>  [966] -2.550711e-02 -4.025946e-02  6.126934e-03  2.834323e-02  9.027484e-03
#>  [971] -5.034417e-02 -1.407961e-02 -7.083999e-02 -3.950338e-02 -1.045910e-02
#>  [976]  4.110606e-02  6.672767e-03 -9.293325e-02 -1.402656e-02  8.378193e-03
#>  [981] -9.598909e-02 -1.888926e-01 -8.193460e-02  6.038278e-03 -9.243886e-02
#>  [986] -1.150398e-01  7.822052e-02  8.403587e-02  4.759178e-02 -3.006136e-03
#>  [991]  6.634066e-02  2.867427e-02  3.032862e-02 -2.404866e-02  5.304512e-02
#>  [996]  1.326244e-02 -4.160260e-02  2.465133e-02  5.386323e-02  1.051685e-02
#> [1001] -8.892806e-02 -6.063962e-02  6.228012e-02 -5.232842e-02  8.036997e-02
#> [1006]  3.193628e-02 -5.922561e-03  5.911278e-02  1.751275e-02  2.653771e-02
#> [1011] -4.840513e-03  2.358457e-02 -1.686195e-02 -2.231743e-02 -2.557481e-02
#> [1016] -6.165341e-02 -7.895598e-02  9.703264e-02 -9.032941e-03  2.975927e-03
#> [1021]  3.765415e-02  3.585724e-02  2.585058e-02 -1.239119e-02 -6.818376e-02
#> [1026]  3.380905e-02  7.481394e-03  1.557511e-02  1.793537e-02 -2.479967e-02
#> [1031] -1.538050e-03  1.339298e-03
#> 
#> $Wald_Joint
#> [1] 1.063572
#> 
#> $Wald_Ind
#>       x1 
#> 1.063572 
#> 
#> $rank
#> [1] 1
#> 
#> $rank_ols
#> [1] 1
#> 
#> $horizon
#> [1] 1
#> 
#> $df.residuals
#> [1] 1031
#> 
#> $df
#> [1] 1
#> 
#> $assign
#> NULL
#> 
#> $cnames
#> [1] "x1"
#> 
#> $AR
#>          Rn        Rz
#> x1 0.999064 0.9986291
#> 
#> $delta
#>            [,1]
#> [1,] -0.1081335
#> 
#> $vcov
#>             [,1]
#> [1,] 0.004156707
#> 
#> $coefficients_ols
#>    Intercept           x1 
#>  0.007972991 -0.062168349 
#> 
#> $tstat_ols
#>           [,1]
#> [1,]  2.174468
#> [2,] -1.011271
#> 
#> $residuals_ols
#>    [1] -0.0109878335  0.0362487054 -0.0012424562  0.0046083367  0.0525424965
#>    [6] -0.0293058632  0.0726443648  0.0245479172  0.0434557629 -0.0562748153
#>   [11]  0.0595548154  0.0105415560 -0.0160792075 -0.0209725343  0.0919818050
#>   [16]  0.0255735256  0.0035390855 -0.0441500364  0.0012663110  0.0586182670
#>   [21]  0.0120360661  0.0048299839  0.1057696444 -0.0031789016  0.0434199934
#>   [26] -0.0130789006 -0.0077821954  0.0097177687 -0.0518987317  0.0948740285
#>   [31]  0.0388476495  0.0818861449 -0.0620805421 -0.2132859978 -0.1429227666
#>   [36]  0.0184252204  0.0488835461  0.0166888169  0.0614585317 -0.0225358739
#>   [41] -0.0194042502 -0.1766722006  0.0304026097  0.0036945876 -0.1388361189
#>   [46] -0.0950427510 -0.0280155773 -0.0815859559  0.0509186539  0.0983150848
#>   [51] -0.0791167563 -0.1042710291 -0.1548646184  0.1301677586 -0.0845241913
#>   [56]  0.0012731681 -0.3450498782  0.0672151069 -0.1146728159 -0.1561905246
#>   [61] -0.0269560891  0.0440957990 -0.1349884351 -0.2400734487 -0.2719950493
#>   [66] -0.0016453811  0.3259534418  0.3327523203 -0.0355805313 -0.1606729619
#>   [71] -0.0594119259  0.0531381014  0.0081437977 -0.2039554680  0.0333502411
#>   [76]  0.3411167887  0.1554333432  0.1236919176 -0.0945384155  0.1113744843
#>   [81] -0.1317113977 -0.1004978267  0.0975033698  0.0216347508  0.1024604359
#>   [86] -0.0400998646 -0.0009891041 -0.0324256428 -0.0884771449  0.0172321493
#>   [91] -0.1340812814  0.0510335850 -0.0067575954 -0.0394125601  0.0768678777
#>   [96] -0.0063604693 -0.0496880725 -0.0433627606 -0.0392062713  0.0854140318
#>  [101]  0.0304153065  0.0615409631  0.0741831135  0.0197941993  0.0197205580
#>  [106]  0.0633447695  0.0445065751  0.0328339578  0.0628332437  0.0166922157
#>  [111]  0.0183701412 -0.0874275105  0.0451595180  0.0257323475  0.0644641916
#>  [116]  0.0075468822 -0.0016203162  0.0672460371  0.0040748624 -0.0103865984
#>  [121]  0.0295876945  0.0120480129 -0.0096620386 -0.0905300517 -0.0122323868
#>  [126] -0.0622443662  0.0877248082 -0.0590035423 -0.1570781241 -0.1153847715
#>  [131] -0.0949536026 -0.0530497454  0.0062762729  0.0575286342 -0.3000200343
#>  [136]  0.1268772599 -0.0474890150  0.2200969356  0.0660408088 -0.0326317911
#>  [141]  0.0104383882  0.0688718364 -0.0327333555  0.0377623437 -0.0798109700
#>  [146]  0.0318982188 -0.1509782309 -0.0134777174  0.0604102545 -0.0706798538
#>  [151]  0.0987822448 -0.0764996347  0.1611824979 -0.0237089737 -0.0477814164
#>  [156]  0.0212096969 -0.0433866797  0.0065056336  0.0069815130 -0.0111973474
#>  [161] -0.2696907661  0.0729661206  0.0260553650  0.0283781958  0.0077857681
#>  [166]  0.0340972365 -0.0360221559 -0.0040451751 -0.0535031505 -0.0112178175
#>  [171]  0.0025104806 -0.0706976348  0.0067198571  0.0504264252  0.0485084547
#>  [176] -0.0081083174 -0.0140536450 -0.0730806791 -0.0317216428 -0.0464958239
#>  [181]  0.0138958624 -0.0250433055 -0.0727171728 -0.0455570015  0.0618853122
#>  [186]  0.0164331143  0.0286954349  0.0083129380  0.0251919443  0.0584268532
#>  [191] -0.0102678002  0.0487705309  0.0628609416  0.0498435031  0.0504958655
#>  [196] -0.0031933956  0.0443928188  0.0149085543 -0.0608319207  0.0091196617
#>  [201]  0.0203933278 -0.0183076813 -0.0755865761  0.0609986630  0.0114711890
#>  [206] -0.0003245059  0.0140794194 -0.0191090587  0.0407700945  0.0459228378
#>  [211] -0.0264904412  0.0075787441 -0.0074142668 -0.0040227822  0.0104422436
#>  [216]  0.0369862432  0.0053817211  0.0602609016 -0.0497998749  0.0798037854
#>  [221]  0.0130294732 -0.0052659133 -0.0284328680  0.0506838995  0.0381475475
#>  [226]  0.0227110745  0.0369552755  0.0031434710  0.0616607754 -0.0746602168
#>  [231]  0.0395443383  0.0283422532  0.0222589484 -0.0431730610 -0.0348170248
#>  [236] -0.0778957022 -0.1151395387 -0.0118080076 -0.0095611854  0.0392830171
#>  [241]  0.0148045306 -0.0153008999 -0.0220819794 -0.0473323111 -0.0079330463
#>  [246]  0.0456390379  0.0318632811 -0.0305398795 -0.0172708813  0.0130529043
#>  [251] -0.0251264052  0.0230059097 -0.0457263262 -0.0483329550  0.0709866913
#>  [256]  0.0230449421  0.0733552772  0.0002658851 -0.0609887845  0.0080871685
#>  [261] -0.0347041575  0.0602466962 -0.1109207314  0.0278860252 -0.0047943719
#>  [266] -0.0396503531  0.0259059411 -0.0264041784 -0.0374018634 -0.0057938131
#>  [271]  0.0532444885  0.0140850060  0.0195112190  0.0217988962  0.0129982154
#>  [276]  0.0421732028  0.0119214895  0.0118866761  0.0005761986  0.0372960600
#>  [281]  0.0382339092 -0.0633645290  0.0092504058  0.0393349109  0.0501229193
#>  [286] -0.0026169138  0.0145168729  0.0462304182  0.0553769226  0.0067261080
#>  [291] -0.0253268942  0.0402995593 -0.0380930241 -0.0338726778  0.0609607306
#>  [296]  0.0412941621 -0.0049288437 -0.0211680634  0.0036472197  0.0333334824
#>  [301]  0.0098707038 -0.0331793751  0.0438748582 -0.0499674719  0.0254412710
#>  [306]  0.0389573045  0.0111698464 -0.0135477519 -0.0260874665 -0.0073549642
#>  [311]  0.0518222207  0.0296804013 -0.0142049889 -0.0154390641 -0.0303692105
#>  [316] -0.0334525952 -0.0005388794 -0.0221880241  0.0181152302 -0.0591998958
#>  [321] -0.0056032343  0.0420878940  0.0161106320 -0.0056827219  0.0445026420
#>  [326]  0.0056865580  0.0230267325  0.0401224549  0.0348035453 -0.0044076646
#>  [331]  0.0502274822 -0.0331542843  0.0722061459 -0.0251580829  0.0843564011
#>  [336]  0.0458213799  0.0106746856  0.0062242142 -0.0093767169  0.0318210999
#>  [341] -0.0010818349  0.0707676581  0.0487199156 -0.0080885954  0.0035389838
#>  [346] -0.0387856859  0.0761660063 -0.0083129987 -0.0439397008  0.0323990868
#>  [351]  0.0598822187 -0.0085997370 -0.0683934693  0.0298852915  0.0419612820
#>  [356] -0.0412878437 -0.0531372209 -0.0017254160 -0.0102642507  0.0272498853
#>  [361] -0.0495695531 -0.0337569174  0.0150053705  0.0368857059  0.0295183253
#>  [366] -0.0091365907  0.0011628589 -0.0584971247 -0.0681545230 -0.0480547817
#>  [371]  0.0146200833 -0.0468578731  0.0363654658 -0.0220141037  0.0253566427
#>  [376]  0.0258406508  0.0141033802  0.0208514003  0.0364476392  0.0111312302
#>  [381]  0.0424560031  0.0197008941  0.0202973246  0.0445471458 -0.0015023661
#>  [386] -0.0010137335 -0.0045423537  0.0315224867  0.0153833589 -0.0086058119
#>  [391]  0.0257085079 -0.0191602485 -0.0521747422  0.0037575147  0.0094861449
#>  [396]  0.0204064107 -0.0815456799  0.0047256234 -0.0212704610 -0.0244786344
#>  [401]  0.0248300551  0.0136909502 -0.0307080916  0.0235194812 -0.0678960849
#>  [406] -0.0088223182  0.0374455771  0.0401344066  0.0547714538  0.0249511176
#>  [411]  0.0213429052 -0.0029389857  0.0157422750 -0.0344496834  0.0257995327
#>  [416]  0.0171409054 -0.0262735062  0.0205382324  0.0367507157 -0.0028040888
#>  [421] -0.0456694548  0.0127319958 -0.0115019634 -0.0715197520 -0.0910938906
#>  [426] -0.0919019105  0.0554870969  0.0135955323 -0.0578750375  0.0004659597
#>  [431]  0.0950003650  0.0075479017  0.0412334749 -0.0324071136  0.0280111632
#>  [436]  0.0400185796  0.0116140446 -0.0266678964 -0.0097539665  0.0439964466
#>  [441] -0.0186558474  0.0259955347 -0.0127609760  0.0174028371  0.0187129691
#>  [446]  0.0085871236  0.0088108119 -0.0021098016  0.0084444296  0.0096771400
#>  [451]  0.0104694885 -0.0199239392  0.0216904150  0.0004139324 -0.0067295620
#>  [456] -0.0031638469  0.0256613136 -0.0047804554 -0.0208948930  0.0263734211
#>  [461] -0.0117460188 -0.0565941584  0.0058771475  0.0186551163  0.0243143590
#>  [466]  0.0198622947 -0.0124621265  0.0016619112 -0.0031299817 -0.0217815792
#>  [471] -0.0321856649  0.0129851288 -0.0593007772 -0.0233760323 -0.0209537032
#>  [476] -0.0833149596 -0.0152591799  0.0377473567  0.0016423363 -0.0090864981
#>  [481]  0.0669647797 -0.0016275561  0.0314384246  0.0337687731 -0.0561434761
#>  [486]  0.0111627721  0.0376405933 -0.0151830725  0.0249262883 -0.0362671879
#>  [491] -0.0019383627  0.0183368872 -0.0514265734 -0.0360672044  0.0009044168
#>  [496]  0.0719521281  0.0071956966  0.0005891991 -0.0268405925  0.0071076049
#>  [501]  0.0296235473 -0.0004371839  0.0431482150 -0.0506684281 -0.0167853195
#>  [506] -0.0524834547  0.0250355640  0.0133929324 -0.0055185631 -0.0649361685
#>  [511] -0.0698733624  0.0347845721 -0.0340884434  0.0361782754 -0.0393151454
#>  [516] -0.0268524290 -0.0885894544  0.0476152741 -0.0070478452 -0.1024003444
#>  [521] -0.0652790658 -0.0593125646  0.0636905702  0.0395160819  0.0251275048
#>  [526] -0.0179056177  0.0443819195  0.0486234123  0.0339416514  0.0053676487
#>  [531]  0.0311829664  0.0291990626 -0.0451939167 -0.0051017176 -0.0496223433
#>  [536]  0.0327110206 -0.0133396712 -0.0491124176 -0.0051334597  0.0764525935
#>  [541]  0.0113104738  0.0227550464  0.0003948017 -0.0018045221  0.0145435481
#>  [546] -0.0276937257 -0.0042126620  0.0310700894 -0.0109007562  0.0027378588
#>  [551]  0.0408319209  0.0052092050 -0.0253591080 -0.0426157167 -0.0094452175
#>  [556] -0.0496590709 -0.0229533478 -0.0140800124  0.0288490740 -0.0419532452
#>  [561]  0.0307059344 -0.0101751845 -0.1231033192  0.0077610220 -0.0178018563
#>  [566] -0.0078366415 -0.0312981308 -0.0480926685 -0.0370820542 -0.0226711097
#>  [571] -0.0892855826 -0.0948792097 -0.1351054807  0.1459259161 -0.0559926468
#>  [576] -0.0275708714  0.1073578674  0.0570544976  0.0160832118  0.0408265570
#>  [581]  0.0425853562  0.0380590886 -0.0751067644 -0.0237015255 -0.0408597889
#>  [586]  0.0548888211  0.0218094299 -0.0178599090  0.1059648284 -0.0128098392
#>  [591]  0.0250731207 -0.0168637416 -0.0152091081  0.0358272760 -0.0146822657
#>  [596] -0.0057195789  0.0167916979 -0.0277600772 -0.0082632364  0.0451118538
#>  [601] -0.0580382484 -0.0228274999 -0.0196198648 -0.0063588043 -0.0221346200
#>  [606]  0.0393278961 -0.0230337467 -0.0199005312 -0.0086754041 -0.0509441120
#>  [611]  0.0278549040 -0.0026584218 -0.0705861355 -0.0246460153  0.0183455414
#>  [616]  0.0771920078  0.0056344865 -0.0247206508  0.0472220150  0.0246239038
#>  [621] -0.0137912363 -0.1029851391  0.0173364376  0.0067952546  0.0318821697
#>  [626] -0.0392937363  0.0457114153 -0.0060600884 -0.0270545195  0.0291965775
#>  [631]  0.0019364970  0.0499473664 -0.0088213529 -0.0785315001  0.0401838760
#>  [636]  0.0066965978  0.0478372223 -0.0083901702 -0.1138660082  0.0286625431
#>  [641]  0.0418478154  0.0206381701  0.0583817555  0.0053739783  0.0190866006
#>  [646]  0.0085243180  0.0943221137 -0.0428160190 -0.0576833237  0.0081606290
#>  [651]  0.0253238565 -0.0317042313 -0.0045182021 -0.0218898626 -0.0113527658
#>  [656] -0.0689792371 -0.0636157903  0.0398672066  0.0323047244 -0.0368295821
#>  [661] -0.0241670645 -0.0611151093 -0.0172559557  0.0308485180 -0.0400694475
#>  [666] -0.0254599281 -0.0299386534  0.1109341574  0.0030569215  0.1006373634
#>  [671]  0.0372738384  0.0090133226  0.0269240494  0.0178251836  0.0267366861
#>  [676]  0.0648148809 -0.0139606984  0.0284413139 -0.0400509662  0.0076760764
#>  [681]  0.0058489754 -0.0214467956  0.0158045301 -0.0137010659 -0.0119090711
#>  [686] -0.0426530660  0.0091935875 -0.0013272140 -0.0626276732  0.0151320627
#>  [691] -0.0214830346  0.0982214086 -0.0088287695 -0.0055742763 -0.0186542897
#>  [696]  0.0173800117  0.0666633238  0.0066541346 -0.0079610673 -0.0114166991
#>  [701]  0.0534982990  0.0083526250 -0.0105157160 -0.0135671277 -0.0410790593
#>  [706]  0.0369468734  0.0620326129  0.0393294536 -0.0032339670  0.0658466195
#>  [711]  0.0457183551 -0.0220954915  0.0457010067  0.0078083224 -0.0674862087
#>  [716]  0.0639880480 -0.0947401295  0.0478335882  0.0172833809 -0.0354770719
#>  [721]  0.1191282160  0.0332777827  0.0181181160 -0.0167642781  0.0017648726
#>  [726]  0.0417157346  0.0414059305  0.0312978710 -0.0293590028 -0.2502098651
#>  [731] -0.0927321745  0.0644017900  0.0347746986  0.0388476347 -0.0387021765
#>  [736]  0.0033507147  0.0008696114  0.0393167070 -0.0121090146 -0.0407057642
#>  [741]  0.0339178134  0.0182484322 -0.0230115726  0.0097022931  0.0608402477
#>  [746] -0.0345768198  0.0135331937  0.0408938870  0.0307545092 -0.0151418500
#>  [751]  0.0765886425  0.0101199933 -0.0133184278 -0.0330616378  0.0109882855
#>  [756]  0.0130664671 -0.0792970543  0.0043503516  0.0172843510 -0.0338172016
#>  [761]  0.0842782967 -0.0156897247 -0.0134866017 -0.1041477232 -0.0586407484
#>  [766] -0.0120088893  0.0542150142  0.0190974806  0.0356909084  0.0611807508
#>  [771]  0.0163797485 -0.0053734619  0.0345672760 -0.0536772756  0.0382186673
#>  [776]  0.0162763125 -0.0243242162  0.0059659883 -0.0482501216  0.1011763244
#>  [781] -0.0253390414  0.0064725588 -0.0259994380  0.0218330796 -0.0012627229
#>  [786] -0.0212821716  0.0338463101 -0.0269614524  0.0052054613 -0.0011965665
#>  [791]  0.0274887163  0.0073635426  0.0011163061  0.0074726747  0.0149080343
#>  [796] -0.0308886854  0.0199163510 -0.0026186115 -0.0109787915  0.0303225881
#>  [801] -0.0142242971  0.0131585436 -0.0153060797  0.0060579112  0.0264284643
#>  [806] -0.0340297999 -0.0508662718  0.0063573506  0.0097740141 -0.0317722106
#>  [811]  0.0255850655  0.0335722349 -0.0311995801  0.0155248531 -0.0439813656
#>  [816]  0.0074958434  0.0180061957  0.0306072221  0.0208545512  0.0209283349
#>  [821]  0.0309983100  0.0153898167  0.0245034884 -0.0046947482  0.0329670780
#>  [826] -0.0116895147  0.0347089191  0.0089325458  0.0257686930  0.0016915174
#>  [831]  0.0016463375  0.0071741482  0.0172428922 -0.0035706244 -0.0532940213
#>  [836]  0.0138308671  0.0468622571  0.0186161872  0.0657220290 -0.0281724215
#>  [841]  0.0530147963 -0.0001503899 -0.0503701697  0.0507749603  0.0517088301
#>  [846]  0.0353436512  0.0695245105 -0.0647509758  0.0444148213 -0.0412031674
#>  [851]  0.0370471859  0.0085716094  0.0034014045  0.0609825375  0.0415799336
#>  [856]  0.0027985488 -0.0261665880  0.0317023362 -0.0186939492 -0.1629292353
#>  [861]  0.0526132537  0.0684706203  0.0523988112  0.0496451223  0.0336542791
#>  [866] -0.0407069546  0.0302216198  0.0288850428 -0.0312832700  0.0451086615
#>  [871] -0.0385621725 -0.0127864162 -0.0363446363  0.0544097201  0.0126405332
#>  [876]  0.0526254263 -0.0589357242 -0.0258805150  0.0853670773 -0.0409560116
#>  [881] -0.0314289283  0.0170735610 -0.0209071367  0.0526263985 -0.0626830083
#>  [886] -0.0133817458 -0.0908734001 -0.0038725673  0.0224420163 -0.1042895167
#>  [891] -0.0742230577  0.0667160339 -0.0004978346 -0.0320123543 -0.0165829674
#>  [896] -0.0729424289 -0.0912194457  0.0125681504  0.0692125346  0.0029170928
#>  [901] -0.0202301883 -0.0254829228  0.0309707910 -0.0686707418 -0.0136611968
#>  [906] -0.0798393416 -0.0827239731  0.0010282606 -0.1215836705  0.0785452269
#>  [911]  0.0512170901 -0.0663942416 -0.0328938258 -0.0212335975  0.0042420653
#>  [916]  0.0736334342  0.0459576547  0.0065808501  0.0118437641  0.0140542662
#>  [921] -0.0162520455  0.0483616636  0.0036983460  0.0450217598  0.0133523901
#>  [926]  0.0086964075 -0.0208697646 -0.0215159446  0.0081629767  0.0138307467
#>  [931] -0.0392098736 -0.0020295339  0.0046182503  0.0085487804  0.0332949771
#>  [936]  0.0266018362 -0.0308997602  0.0139084760 -0.0244998956 -0.0263272947
#>  [941]  0.0240718360 -0.0058859262  0.0289146093 -0.0170190551 -0.0001807219
#>  [946] -0.0236085267  0.0295964209 -0.0082708122  0.0177394117 -0.0061302547
#>  [951]  0.0040096228  0.0034125496 -0.0372183278 -0.0070158144 -0.0031209169
#>  [956]  0.0150133428  0.0171280379  0.0230322412  0.0091437282  0.0044872592
#>  [961]  0.0062793349 -0.0284506190  0.0016465998  0.0340345613  0.0248955059
#>  [966] -0.0255000529 -0.0402571551  0.0061378841  0.0283602339  0.0090436204
#>  [971] -0.0503241471 -0.0140444618 -0.0708070021 -0.0394643362 -0.0104209180
#>  [976]  0.0411468296  0.0067023020 -0.0929110660 -0.0139978882  0.0084047025
#>  [981] -0.0959556654 -0.1888565834 -0.0819137064  0.0061049813 -0.0923423368
#>  [986] -0.1149825732  0.0782746862  0.0841099206  0.0476420637 -0.0029653645
#>  [991]  0.0663827278  0.0287159098  0.0303767365 -0.0239953540  0.0530910800
#>  [996]  0.0133144496 -0.0415730650  0.0246883887  0.0539002353  0.0105464844
#> [1001] -0.0888896010 -0.0605879663  0.0623450593 -0.0522636963  0.0804560855
#> [1006]  0.0320165453 -0.0058537652  0.0591759185  0.0175613918  0.0265784839
#> [1011] -0.0047971486  0.0236266382 -0.0168142648 -0.0222589381 -0.0255219346
#> [1016] -0.0615841103 -0.0788642106  0.0971455849 -0.0089299337  0.0030862815
#> [1021]  0.0377744463  0.0359771072  0.0259618032 -0.0122890511 -0.0680690856
#> [1026]  0.0339418748  0.0076116293  0.0157135588  0.0180686323 -0.0246698630
#> [1031] -0.0014125691  0.0014721267
#>