Fitting IVX Models

ivx fits predictive regression models. The method allows standard chi-square testing for regressors with different degrees of persistence, from stationary to mildly explosive, and can be used for both short- and long-horizon predictive regressions.

ivx(
  formula,
  data,
  horizon,
  na.action,
  weights,
  contrasts = NULL,
  offset,
  model = TRUE,
  x = FALSE,
  y = FALSE,
  ...
)

# S3 method for class 'ivx'
print(x, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

formula

an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data

n optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.

horizon

is the horizon (default horizon = 1 corresponds to a short-horizon regression).

na.action

a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful.

weights

an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector. If non-NULL, weighted least squares is used with weights weights (that is, minimizing sum(w*e^2)); otherwise ordinary least squares is used.

contrasts

an optional list. See the contrasts.arg of model.matrix.default.

offset

this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector or matrix of extents matching those of the response. One or more offset terms can be included in the formula instead or as well, and if more than one are specified their sum is used. See model.offset

model

logical. If TRUE the model.frame of the fit is returned.

x

an object of class "ivx", usually, a result of a call to ivx.

y

logical. If TRUE the response of the fit is returned.

...

additional arguments to be passed to the low level regression fitting functions (see lm).

digits

the number of significant digits to use when printing.

Value

an object of class "ivx".

References

Magdalinos, T., & Phillips, P. (2009). Limit Theory for Cointegrated Systems with Moderately Integrated and Moderately Explosive Regressors. Econometric Theory, 25(2), 482-526.

Kostakis, A., Magdalinos, T., & Stamatogiannis, M. P. (2014). Robust econometric inference for stock return predictability. The Review of Financial Studies, 28(5), 1506-1553.

Examples

# Univariate
ivx(Ret ~ LTY, data = kms)
#> 
#> Call:
#> ivx(formula = Ret ~ LTY, data = kms, horizon = 1)
#> 
#> Coefficients:
#>      LTY  
#> -0.06649  
#> 

# Multivariate
ivx(Ret ~ LTY + TBL, data = kms)
#> 
#> Call:
#> ivx(formula = Ret ~ LTY + TBL, data = kms, horizon = 1)
#> 
#> Coefficients:
#>      LTY       TBL  
#>  0.07624  -0.13497  
#> 

# Longer horizon
ivx(Ret ~ LTY + TBL, data = kms, horizon = 4)
#> 
#> Call:
#> ivx(formula = Ret ~ LTY + TBL, data = kms, horizon = 4)
#> 
#> Coefficients:
#>      LTY       TBL  
#>  0.09322  -0.14164  
#> 

wt <- runif(nrow(kms))
ivx(Ret ~ LTY, data = kms, weights = wt)
#> 
#> Call:
#> ivx(formula = Ret ~ LTY, data = kms, weights = wt, horizon = 1)
#> 
#> Coefficients:
#>      LTY  
#> -0.05046  
#>